Docs: Configure meters/events dispatch separately
This is a docs change in support of Change-Id:
I161318f4503ce79205b54a93f2ae9b4e052f7971

Dispatchers now only record meter or event data and are configured
independently. The install guide is updated to reflect this
distinction. Because the referenced Change-Id used Doc-Impact
as a flag instead of the non-hyphenated variant there is no
independant bug in the OpenStack Manuals project for this change.

Additionally, existing content for configuring Gnocchi is broken out
into more detail.

Change-Id: I11e9a3037e8be6d0d76c41f4ddd4e94583dc132f
Related-Bug: #1480333