Cache getters for the decalarative definitions
In the our declarative configs for the events and meters we use
same fields for the many times. Despite this fact all fields are
parsed by ExtentedJsonPathParser again and again.
It affect time of notification agent starting and create additional
cpu load.

This patch is a trying to make this process faster and lighter.

Change-Id: Ib190b5977688cdf48cc454ff133dd81ff5576310
Closes-bug: #1550436