timedelta plugin for meter definition process
with timedelta plugin we could create new metric for latency time of
nstance booting without using events and its transformation.
we could define new meter in meters.yaml with volume as in example::
   volume:
      fields: [$.payload.created_at, $.payload.launched_at]
      plugin: ‘timedelta’

as a result we get volume value equal to difference between two mentioned
timestamp fields in seconds.

Change-Id: If5084cc23212a0a6bd9dac8438d5d286f3415730